refactor launch bundle tests
